+++ tests/testConfig.js
@@ -6,18 +6,20 @@
  */
 
 const TEST_FILES = [
-  "testharness.test.js",
-  "../js/constants.test.js",
-  "../js/utils.test.js",
-  "../js/solver.test.js",
-  "../js/state.test.js",
-  "../js/puzzles.test.js",
-  "../js/render.test.js",
-  "../js/theme.test.js",
-  "../js/dom.test.js",
-  "../js/app.test.js",
+  'testharness.test.js',
+  '../js/constants.test.js',
+  '../js/utils.test.js',
+  '../js/solver.test.js',
+  '../js/styler.test.js',
+  '../js/state.test.js',
+  '../js/puzzles.test.js',
+  '../js/render.test.js',
+  '../js/theme.test.js',
+  '../js/dom.test.js',
+  '../js/app.test.js',
+  '../js/options.test.js',
 ];
 
-if (typeof module !== "undefined" && module.exports) {
+if (typeof module !== 'undefined' && module.exports) {
   module.exports = { TEST_FILES };
 }